\name{area_circle}
\alias{area_circle}
\title{Area of a circle}
\description{Compute the area of a circle by providing the radius.}
\usage{area_circle(radius)}
\arguments{
  \item{radius}{numeric. Cannot be negative}
}
\value{numeric}

\examples{
area_circle(10)
}

% the below is not required 
\details{this is a very smart function!}
\references{Pythagoras, -2000}
\author{Robert Hijmans}
\note{developed during the workshop}
\section{Warning }{not much tested yet!}
\seealso{\code{\link{length}}}

\keyword{spatial}
\concept{workshop}
