% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCDC.R
\name{filter.CDC.data}
\alias{filter.CDC.data}
\title{This function the CDC dataset as input and apply year and state filters}
\usage{
filter.CDC.data(ds, year, state)
}
\arguments{
\item{ds}{CDC dataset,}

\item{year}{year to filter data}

\item{state}{US State Which state should the filter be applied}
}
\value{
filtered dataset
}
\description{
Takes in CDC dataset, year, state as input paramaters.  Also remove diseases "All Causes"
}
