% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCDC.R
\name{get.CDC.data}
\alias{get.CDC.data}
\title{Get the dataset from CDC website that contains the disease which is the leading cause of death}
\usage{
get.CDC.data()
}
\value{
Complete dataset from CDC website
}
\description{
Takes in no paramaters, get the data and rename the columns
}
