% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIMDBData.R
\name{get.IMDB.data}
\alias{get.IMDB.data}
\title{Get web page data from IMDB website}
\usage{
get.IMDB.data(url)
}
\arguments{
\item{url}{Url of web page}
}
\value{
Complete web page from IMDB website
}
\description{
Takes in the URL as input paramater, return the entire web page
}
