% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopular.R
\name{getpopular}
\alias{getpopular}
\title{Scrapes data from IMDB website, remove unwanted data and relevant attributes in a dataset}
\usage{
getpopular(p1)
}
\arguments{
\item{p1}{Input parameter valid value are 'movies' 'tvshows'}
}
\value{
Complete formatted dataset extracted from IMDB website
}
\description{
Takes in no paramaters, get the data and rename the columns
}
