% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTop.R
\name{gettopshows}
\alias{gettopshows}
\title{Scrapes data from IMDB website, remove unwanted data and package relevant data in a dataset}
\usage{
gettopshows(p1)
}
\arguments{
\item{p1}{Input parameter valid value are 'movies' 'tvshows'}
}
\value{
Formatted dataset of Top 100 movies or tvshows
}
\description{
Takes in no paramaters, get the data and rename the columns
}
