% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCols.R
\name{parseCols.data}
\alias{parseCols.data}
\title{Private function that parses individual columns}
\usage{
parseCols.data(data, str)
}
\arguments{
\item{data}{css Tag where data is extracted}

\item{str}{regular expression string}
}
\value{
vector of specific attribute from IMDB data
}
\description{
Takes in unformatted datas
}
