% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCDC.R
\name{plot.CDC.data}
\alias{plot.CDC.data}
\title{This function plots a barchart}
\usage{
\method{plot}{CDC.data}(ds)
}
\arguments{
\item{ds}{CDC dataset,}
}
\value{
null
}
\description{
Takes in filtered CDC dataset
}
