\name{R2D2-package}
\alias{R2D2-package}
\alias{R2D2}
\docType{package}
\title{
\packageTitle{R2D2}
}
\description{
\packageDescription{R2D2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{R2D2}
\packageIndices{R2D2}
{Bayesian Linear Regression Using the R2-D2 Shrinkage Prior
}
\author{
\packageAuthor{R2D2}
Maintainer: \packageMaintainer{R2D2}
}
\references{
Zhang, Yan Dora, et al. "Bayesian Regression Using a Prior on the Model Fit: The R2-D2 Shrinkage Prior." arXiv:1609.00046v3 (2020)
}
\keyword{Global-local shrinkage, High-dimensional regression, Beta-prime distribution, Coefficient of Determination}
