% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinvgauss_new.R
\name{rinvgauss_new}
\alias{rinvgauss_new}
\title{The Inverse Gaussian Distribution}
\usage{
rinvgauss_new(n, mean = 1, shape = NULL, dispersion = 1)
}
\arguments{
\item{n}{Sample size.}

\item{mean}{Vector of means. Default to be 1.}

\item{shape}{Vector of shape parameters.}

\item{dispersion}{Vector of dispersion parameters.}
}
\description{
\code{rinvgauss_new} achieves random generation from the inverse gaussian
distribution.
}
\keyword{Internal}
