\name{Ushape}
\alias{Ushape}
\docType{data}
\title{U-Shaped Cloud}
\description{
  Bivariate scatter shaped like an open circle, for testing spatial
  algorithms.
}
\usage{Ushape}
\format{
  Matrix containing 1000 rows and 2 columns:
  \tabular{ll}{
    \code{x}\tab x coordinates.\cr
    \code{y}\tab y coordinates.
  }
}
\examples{
freq2d(Ushape)
conf2d(Ushape)
}
\keyword{datasets}
